refactor(cli): extract workflow history handler
