+++ application/src/Nexo.CLI/Commands/WorkflowCommand.cs
@@ -41,6 +41,7 @@ public sealed class WorkflowCommand : Command
     private readonly Func<IReadOnlyList<string>, CancellationToken, Task<ModelPullResult>> _ollamaModelPuller;
     private readonly Func<CancellationToken, Task<IReadOnlyList<PeerInfo>>> _meshPeerDiscovery;
     private readonly ScaffoldHandler _scaffoldHandler;
+    private readonly HistoryHandler _historyHandler;
 
     public WorkflowCommand(Func<OrchestrateCommand> orchestrateFactory)
         : this(
@@ -172,6 +173,7 @@ public sealed class WorkflowCommand : Command
         _meshPeerDiscovery = meshPeerDiscovery ?? (_ => Task.FromResult<IReadOnlyList<PeerInfo>>(Array.Empty<PeerInfo>()));
         _meshScenarioExecutor = meshScenarioExecutor ?? ExecuteScenarioOnMeshPeerAsync;
         _scaffoldHandler = new ScaffoldHandler();
+        _historyHandler = new HistoryHandler();
         ConfigureScaffoldCommand();
         ConfigureStressCommand();
         ConfigureHistoryCommand();
@@ -645,39 +647,7 @@ public sealed class WorkflowCommand : Command
 
     internal Task<int> ExecuteScaffoldAsync(string outputPath, bool force, bool json) => _scaffoldHandler.ExecuteAsync(outputPath, force, json);
 
-    internal Task<int> ExecuteHistoryAsync(string repoRoot, int limit, string? benchmarkSet, bool json)
-    {
-        var fullRepoRoot = Path.GetFullPath(string.IsNullOrWhiteSpace(repoRoot) ? Environment.CurrentDirectory : repoRoot);
-        if (!Directory.Exists(fullRepoRoot))
-        {
-            WriteHistoryResult(new WorkflowHistoryResult(false, $"Repo root not found: {fullRepoRoot}"), json);
-            return Task.FromResult(1);
-        }
-
-        var rows = WorkflowLabHistoryStore.ReadRecent(fullRepoRoot, Math.Max(1, limit));
-        if (!string.IsNullOrWhiteSpace(benchmarkSet))
-        {
-            var normalized = benchmarkSet.Trim().ToLowerInvariant();
-            rows = rows
-                .Where(x => string.Equals(x.BenchmarkSet, normalized, StringComparison.OrdinalIgnoreCase))
-                .ToArray();
-        }
-
-        var total = rows.Count;
-        var successCount = rows.Count(x => x.Success);
-        var best = rows
-            .OrderByDescending(x => x.Success)
-            .ThenByDescending(x => x.Score)
-            .ThenBy(x => x.ElapsedMs)
-            .FirstOrDefault();
-
-        WriteHistoryResult(new WorkflowHistoryResult(
-            true,
-            $"Loaded {total} workflow stress history entries.",
-            rows,
-            new WorkflowHistorySummary(total, successCount, total - successCount, best?.ScenarioId, best?.Score)), json);
-        return Task.FromResult(0);
-    }
+    internal Task<int> ExecuteHistoryAsync(string repoRoot, int limit, string? benchmarkSet, bool json) => _historyHandler.ExecuteAsync(repoRoot, limit, benchmarkSet, json);
 
     internal Task<int> ExecuteBaselinePromoteAsync(
         string repoRoot,
@@ -3582,29 +3552,6 @@ public sealed class WorkflowCommand : Command
         }
     }
 
-    private static void WriteHistoryResult(WorkflowHistoryResult result, bool json)
-    {
-        if (json)
-        {
-            Console.WriteLine(JsonSerializer.Serialize(new
-            {
-                ok = result.Ok,
-                summary = result.Summary,
-                summaryStats = result.SummaryStats,
-                items = result.Items
-            }, new JsonSerializerOptions { WriteIndented = true }));
-            return;
-        }
-
-        Console.WriteLine($"workflow history: {(result.Ok ? "ok" : "failed")}");
-        Console.WriteLine(result.Summary);
-        if (result.SummaryStats != null)
-        {
-            Console.WriteLine(
-                $"  total={result.SummaryStats.Total}, success={result.SummaryStats.Success}, failed={result.SummaryStats.Failed}, best={result.SummaryStats.BestScenarioId ?? "n/a"}, best-score={result.SummaryStats.BestScore?.ToString("F2") ?? "n/a"}");
-        }
-    }
-
     private static void WriteReportResult(WorkflowReportResult result, bool json)
     {
         if (json)
@@ -3913,19 +3860,6 @@ public sealed class WorkflowCommand : Command
         IReadOnlyList<OptimizeAllocationTrace>? AllocationTrace = null,
         IReadOnlyList<TargetAllocationStat>? TargetAllocations = null,
         IReadOnlyList<CandidateAllocationStat>? CandidateAllocations = null);
-
-    private sealed record WorkflowHistorySummary(
-        int Total,
-        int Success,
-        int Failed,
-        string? BestScenarioId,
-        double? BestScore);
-
-    private sealed record WorkflowHistoryResult(
-        bool Ok,
-        string Summary,
-        IReadOnlyList<WorkflowLabStressHistoryRow>? Items = null,
-        WorkflowHistorySummary? SummaryStats = null);
 
     private sealed record WorkflowScenarioBenchmark(
         string ScenarioGroupId,
